requirements.txt
apache-airflow
pandas
requests
SQLAlchemy